+++ packages/xo-server-netbox/src/index.js
@@ -1,7 +1,7 @@
 import assert from 'assert'
 import ipaddr from 'ipaddr.js'
 import { createLogger } from '@xen-orchestra/log'
-import { find, flatten, forEach, groupBy, isEmpty, keyBy, mapValues, trimEnd, zipObject } from 'lodash'
+import { find, flatten, forEach, groupBy, isEmpty, keyBy, mapValues, omit, trimEnd, zipObject } from 'lodash'
 
 const log = createLogger('xo:netbox')
 
@@ -40,6 +40,7 @@ const onRequest = req => {
 }
 
 class Netbox {
+  #allowUnauthorized
   #endpoint
   #intervalToken
   #loaded
@@ -58,6 +59,7 @@ class Netbox {
     if (!/^https?:\/\//.test(this.#endpoint)) {
       this.#endpoint = 'http://' + this.#endpoint
     }
+    this.#allowUnauthorized = configuration.allowUnauthorized ?? false
     this.#token = configuration.token
     this.#pools = configuration.pools
     this.#syncInterval = configuration.syncInterval && configuration.syncInterval * 60 * 60 * 1e3
@@ -105,6 +107,7 @@ class Netbox {
       headers: { 'Content-Type': 'application/json', Authorization: `Token ${this.#token}` },
       method,
       onRequest,
+      rejectUnauthorized: !this.#allowUnauthorized,
     }
 
     const httpRequest = async () => {
@@ -164,7 +167,23 @@ class Netbox {
     return results
   }
 
+  async #checkCustomFields() {
+    const customFields = await this.#makeRequest('/extras/custom-fields/', 'GET')
+    const uuidCustomField = customFields.find(field => field.name === 'uuid')
+    if (uuidCustomField === undefined) {
+      throw new Error('UUID custom field was not found. Please create it manually from your Netbox interface.')
+    }
+    const { content_types: types } = uuidCustomField
+    if (!types.includes('virtualization.cluster') || !types.includes('virtualization.virtualmachine')) {
+      throw new Error(
+        'UUID custom field must be assigned to types virtualization.cluster and virtualization.virtualmachine'
+      )
+    }
+  }
+
   async #synchronize(pools = this.#pools) {
+    await this.#checkCustomFields()
+
     const xo = this.#xo
     log.debug('synchronizing')
     // Cluster type
@@ -215,6 +234,10 @@ class Netbox {
       }
     }
 
+    // FIXME: Should we deduplicate cluster names even though it also fails when
+    // a cluster within another cluster type has the same name?
+    // FIXME: Should we delete clusters from this cluster type that don't have a
+    // UUID?
     Object.assign(
       clusters,
       keyBy(
@@ -234,30 +257,54 @@ class Netbox {
 
     // VMs
     const vms = xo.getObjects({ filter: object => object.type === 'VM' && pools.includes(object.$pool) })
-    const oldNetboxVms = keyBy(
-      flatten(
-        // FIXME: It should be doable with one request:
-        // `cluster_id=1&cluster_id=2` but it doesn't work
-        // https://netbox.readthedocs.io/en/stable/rest-api/filtering/#filtering-objects
-        await Promise.all(
-          pools.map(poolId =>
-            this.#makeRequest(`/virtualization/virtual-machines/?cluster_id=${clusters[poolId].id}`, 'GET')
-          )
+    let oldNetboxVms = flatten(
+      // FIXME: It should be doable with one request:
+      // `cluster_id=1&cluster_id=2` but it doesn't work
+      // https://netbox.readthedocs.io/en/stable/rest-api/filtering/#filtering-objects
+      await Promise.all(
+        pools.map(poolId =>
+          this.#makeRequest(`/virtualization/virtual-machines/?cluster_id=${clusters[poolId].id}`, 'GET')
         )
-      ),
-      'custom_fields.uuid'
+      )
     )
+
+    const vmsWithNoUuid = oldNetboxVms.filter(vm => vm.custom_fields.uuid === null)
+    oldNetboxVms = omit(keyBy(oldNetboxVms, 'custom_fields.uuid'), null)
+
+    // Delete VMs that don't have a UUID custom field. This can happen if they
+    // were created manually or if the custom field config was changed after
+    // their creation
+    if (vmsWithNoUuid !== undefined) {
+      log.warn(`Found ${vmsWithNoUuid.length} VMs with no UUID. Deleting them.`)
+      await this.#makeRequest(
+        '/virtualization/virtual-machines/',
+        'DELETE',
+        vmsWithNoUuid.map(vm => ({ id: vm.id }))
+      )
+    }
 
     // Build collections for later
     const netboxVms = {} // VM UUID → Netbox VM
-    const vifsByVm = {} // VM UUID → VIF
+    const vifsByVm = {} // VM UUID → VIF UUID[]
     const ipsByDeviceByVm = {} // VM UUID → (VIF device → IP)
+    const primaryIpsByVm = {} // VM UUID → { ipv4, ipv6 }
 
     const vmsToCreate = []
-    const vmsToUpdate = []
+    let vmsToUpdate = [] // will be reused for primary IPs
     for (const vm of Object.values(vms)) {
       vifsByVm[vm.uuid] = vm.VIFs
       const vmIpsByDevice = (ipsByDeviceByVm[vm.uuid] = {})
+
+      if (primaryIpsByVm[vm.uuid] === undefined) {
+        primaryIpsByVm[vm.uuid] = {}
+      }
+      if (vm.addresses['0/ipv4/0'] !== undefined) {
+        primaryIpsByVm[vm.uuid].ipv4 = vm.addresses['0/ipv4/0']
+      }
+      if (vm.addresses['0/ipv6/0'] !== undefined) {
+        primaryIpsByVm[vm.uuid].ipv6 = ipaddr.parse(vm.addresses['0/ipv6/0']).toString()
+      }
+
       forEach(vm.addresses, (address, key) => {
         const device = key.split('/')[0]
         if (vmIpsByDevice[device] === undefined) {
@@ -463,7 +510,7 @@ class Netbox {
       .forEach(newInterfaces => Object.assign(interfaces, newInterfaces))
 
     // IPs
-    const [oldNetboxIps, prefixes] = await Promise.all([
+    const [oldNetboxIps, netboxPrefixes] = await Promise.all([
       this.#makeRequest('/ipam/ip-addresses/', 'GET').then(addresses =>
         groupBy(
           // In Netbox, a device interface and a VM interface can have the same
@@ -480,6 +527,7 @@ class Netbox {
     const ipsToDelete = []
     const ipsToCreate = []
     const ignoredIps = []
+    const netboxIpsByVif = {}
     for (const [vmUuid, vifs] of Object.entries(vifsByVm)) {
       const vmIpsByDevice = ipsByDeviceByVm[vmUuid]
       if (vmIpsByDevice === undefined) {
@@ -492,6 +540,8 @@ class Netbox {
           continue
         }
 
+        netboxIpsByVif[vifId] = []
+
         const interface_ = interfaces[vif.uuid]
         const interfaceOldIps = oldNetboxIps[interface_.id] ?? []
 
@@ -499,28 +549,36 @@ class Netbox {
           const parsedIp = ipaddr.parse(ip)
           const ipKind = parsedIp.kind()
           const ipCompactNotation = parsedIp.toString()
-          // FIXME: Should we compare the IPs with their range? ie: can 2 IPs
-          // look identical but belong to 2 different ranges?
-          const netboxIpIndex = interfaceOldIps.findIndex(
-            netboxIp => ipaddr.parse(netboxIp.address.split('/')[0]).toString() === ipCompactNotation
-          )
+
+          let smallestPrefix
+          let highestBits = 0
+          netboxPrefixes.forEach(({ prefix }) => {
+            const [range, bits] = prefix.split('/')
+            const parsedRange = ipaddr.parse(range)
+            if (parsedRange.kind() === ipKind && parsedIp.match(parsedRange, bits) && bits > highestBits) {
+              smallestPrefix = prefix
+              highestBits = bits
+            }
+          })
+          if (smallestPrefix === undefined) {
+            ignoredIps.push(ip)
+            continue
+          }
+
+          const netboxIpIndex = interfaceOldIps.findIndex(netboxIp => {
+            const [ip, bits] = netboxIp.address.split('/')
+            return ipaddr.parse(ip).toString() === ipCompactNotation && bits === highestBits
+          })
+
           if (netboxIpIndex >= 0) {
+            netboxIpsByVif[vifId].push(interfaceOldIps[netboxIpIndex])
             interfaceOldIps.splice(netboxIpIndex, 1)
           } else {
-            const prefix = prefixes.find(({ prefix }) => {
-              const [range, bits] = prefix.split('/')
-              const parsedRange = ipaddr.parse(range)
-              return parsedRange.kind() === ipKind && parsedIp.match(parsedRange, bits)
-            })
-            if (prefix === undefined) {
-              ignoredIps.push(ip)
-              continue
-            }
-
             ipsToCreate.push({
-              address: `${ip}/${prefix.prefix.split('/')[1]}`,
+              address: `${ip}/${smallestPrefix.split('/')[1]}`,
               assigned_object_type: 'virtualization.vminterface',
               assigned_object_id: interface_.id,
+              vifId, // needed to populate netboxIpsByVif with newly created IPs
             })
           }
         }
@@ -534,8 +592,60 @@ class Netbox {
 
     await Promise.all([
       ipsToDelete.length !== 0 && this.#makeRequest('/ipam/ip-addresses/', 'DELETE', ipsToDelete),
-      ipsToCreate.length !== 0 && this.#makeRequest('/ipam/ip-addresses/', 'POST', ipsToCreate),
+      ipsToCreate.length !== 0 &&
+        this.#makeRequest(
+          '/ipam/ip-addresses/',
+          'POST',
+          ipsToCreate.map(ip => omit(ip, 'vifId'))
+        ).then(newNetboxIps => {
+          newNetboxIps.forEach((newNetboxIp, i) => {
+            const { vifId } = ipsToCreate[i]
+            if (netboxIpsByVif[vifId] === undefined) {
+              netboxIpsByVif[vifId] = []
+            }
+            netboxIpsByVif[vifId].push(newNetboxIp)
+          })
+        }),
     ])
+
+    // Primary IPs
+    vmsToUpdate = []
+    Object.entries(netboxVms).forEach(([vmId, netboxVm]) => {
+      if (netboxVm.primary_ip4 !== null && netboxVm.primary_ip6 !== null) {
+        return
+      }
+      const newNetboxVm = { id: netboxVm.id }
+      const vifs = vifsByVm[vmId]
+      vifs.forEach(vifId => {
+        const netboxIps = netboxIpsByVif[vifId]
+        const vmMainIps = primaryIpsByVm[vmId]
+
+        netboxIps?.forEach(netboxIp => {
+          const address = netboxIp.address.split('/')[0]
+          if (
+            newNetboxVm.primary_ip4 === undefined &&
+            address === vmMainIps.ipv4 &&
+            netboxVm.primary_ip4?.address !== netboxIp.address
+          ) {
+            newNetboxVm.primary_ip4 = netboxIp.id
+          }
+          if (
+            newNetboxVm.primary_ip6 === undefined &&
+            address === vmMainIps.ipv6 &&
+            netboxVm.primary_ip6?.address !== netboxIp.address
+          ) {
+            newNetboxVm.primary_ip6 = netboxIp.id
+          }
+        })
+      })
+      if (newNetboxVm.primary_ip4 !== undefined || newNetboxVm.primary_ip6 !== undefined) {
+        vmsToUpdate.push(newNetboxVm)
+      }
+    })
+
+    if (vmsToUpdate.length > 0) {
+      await this.#makeRequest('/virtualization/virtual-machines/', 'PATCH', vmsToUpdate)
+    }
 
     log.debug('synchronized')
   }
@@ -554,6 +664,8 @@ class Netbox {
       'GET'
     )
 
+    await this.#checkCustomFields()
+
     if (clusterTypes.length !== 1) {
       throw new Error('Could not properly write and read Netbox')
     }
@@ -571,6 +683,11 @@ export const configurationSchema = ({ xo: { apiMethods } }) => ({
       type: 'string',
       title: 'Endpoint',
       description: 'Netbox URI',
+    },
+    allowUnauthorized: {
+      type: 'boolean',
+      title: 'Unauthorized certificates',
+      description: 'Enable this if your Netbox instance uses a self-signed SSL certificate',
     },
     token: {
       type: 'string',